test_requirements.txt
pytest
pytest-docker-compose
kafka-python